Feat(variables-scss): Add debug option for exporter #DS-1505
