+++ exporters/variables-scss/config.local.json
@@ -1,3 +1,3 @@
 {
-  "generateDisclaimer": true
+  "generateOriginalDataFiles": false
 }
